(module): Separate methods and add tests for Serve
